+++ objsamples/sample-campaign.php
@@ -1,11 +1,8 @@
 <?php
-// include_once('src/ET_Client.php');
-// include_once('src/ET_Campaign.php');
-// include_once('src/ET_Campaign_Asset.php');
-spl_autoload_register( function($class_name) {
-    include_once 'src/'.$class_name.'.php';
-});
-date_default_timezone_set('UTC');
+include_once('tests/UnitBootstrap.php');
+use FuelSdk\ET_Client;
+use FuelSdk\ET_Campaign;
+use FuelSdk\ET_Campaign_Asset;
 
 try {	
 
@@ -26,7 +23,7 @@ try {
 	print_r('More Results: '.($getResult->moreResults ? 'true' : 'false')."\n");
 	print 'Results Length(Items): '. count($getResult->results->items)."\n";
 	//print 'Results: "\n"';
-	//print_r($getResult->results);
+	print_r($getResult->results);
 	print "\n---------------\n";
 	
 	while ($getResult->moreResults) {