+++ objsamples/sample-resultmessage.php
@@ -1,8 +1,8 @@
 <?php
-spl_autoload_register( function($class_name) {
-    include_once 'src/'.$class_name.'.php';
-});
-date_default_timezone_set('UTC');
+include_once('tests/UnitBootstrap.php');
+use FuelSdk\ET_Client;
+use FuelSdk\ET_ResultMessage;
+
 
 try {	
 	$myclient = new ET_Client();