+++ objsamples/sample-tenant.php
@@ -1,10 +1,7 @@
 <?php
-// include_once('src/ET_Client.php');
-// include_once('src/ET_OEM_Client.php');
-spl_autoload_register( function($class_name) {
-    include_once 'src/'.$class_name.'.php';
-});
-date_default_timezone_set('UTC');
+include_once('tests/UnitBootstrap.php');
+use FuelSdk\ET_Client;
+use FuelSdk\ET_OEM_Client;
 
 try {