+++ src/ET_PatchRest.php
@@ -1,7 +1,8 @@
 <?php
-spl_autoload_register( function($class_name) {
-    include_once 'src/'.$class_name.'.php';
-});
+// spl_autoload_register( function($class_name) {
+//     include_once 'src/'.$class_name.'.php';
+// });
+namespace FuelSdk;
 
 /**
  * This class represents the PATCH operation for REST service.
@@ -14,9 +15,10 @@ class ET_PatchRest extends ET_Constructor
 	* @param 	string 		$url 		The endpoint URL
 	* @param 	array       $props 		Dictionary type array which may hold e.g. array('id' => '', 'key' => '')
 	*/
-	function __construct($authStub, $url, $props)
+	function __construct($authStub, $url, $props, $qs="")
 	{
-		$restResponse = ET_Util::restPatch($url, json_encode($props), $authStub);			
+//		$restResponse = ET_Util::restPatch($url, json_encode($props), $authStub);			
+		$restResponse = ET_Util::restPatch($url, json_encode($props), $authStub, $qs);			
 		parent::__construct($restResponse->body, $restResponse->httpcode, true);							
 	}
 }